+++ CHANGELOG.md
@@ -5,6 +5,17 @@
 - Added default route configuration commands:
   - Configuration -> `ip.setDefaultRoute`
   - Configuration -> `ipv6.setDefaultRoute`
+- Removed Configuration -> `ip.dhcp.createPool`
+- Added IPv6 DHCP pool configuration commands:
+  - Configuration -> `ipv6.dhcp.pool` (to Ipv6 Pool Scope)
+  - Interface -> `ipv6.enable`
+  - Interface -> `ipv6.enableDhcpConsumer`
+  - Interface -> `ipv6.enableDhcpProvider`
+  - Interface -> `ipv6.forceStatefulDhcp`
+  - Interface -> `ipv6.setDhcpServer`
+  - IPv6 Pool -> `setAddressPrefix`
+  - IPv6 Pool -> `setDomainName`
+  - IPv6 Pool -> `setDnsServer`
 
 ## 1.1.3